move GithubFile into everypolitician-extensions
We hopefully shouldn't even _need_ this class for much longer (as we'll
be able to use the built-in in `everypolitician-ruby`) but whilst we
do, it should live in `everypolitician-extensions` rather than
`popolo_helper`.
