+++ lib/popolo_helper.rb
@@ -1,28 +1,4 @@
 # frozen_string_literal: true
-require 'open-uri'
-require 'fileutils'
-
-module EveryPolitician
-  class GithubFile
-    GH_PATH = 'https://cdn.rawgit.com/everypolitician/everypolitician-data/%s/%s'
-
-    def initialize(file, sha, cache_dir = '_cached_data')
-      @url = GH_PATH % [sha, file]
-
-      FileUtils.mkpath cache_dir
-      @cache_file = File.join cache_dir, sha + '-' + file.tr('/', '-')
-    end
-
-    attr_reader :url
-
-    def raw
-      @_data ||= begin
-        File.write(@cache_file, open(@url).read) unless File.exist? @cache_file
-        File.read(@cache_file)
-      end
-    end
-  end
-end
 
 module Popolo
   module Helper
